Create Board.java
The specification for the Board class is as follows:

Variables
private String[][] squares - An array which represents the board on which a player places their battleship and records shots.

Constructors
public Board() - This constructor initializes the squares array with every value set to "-", which is the String used to represent a blank square.

Methods
public String toString() - Returns a multi-line representation of the board by concatenating all the values in squares with a new line for each row and spaces separating the Strings in each row.
public boolean addShip(int row, int col, int len, boolean horizontal) - Attempts to add a ship of length len to the grid, starting at the row and column specified and proceeding either rightwards (if horizontal is true), or downwards (if horizontal is false). If the ship can be placed in the place specified, each square making up the ship should be set to "b", and the method should return true. A ship may not be placed if it would go off the grid, or would intersect another ship on the grid. If the ship cannot be placed, no values in squares should be changed and the method should return false.
public boolean foundShip(int len) - Search the board for any possible ships of length len. If there are exactly len consecutive squares (either horizontal or vertical) containing a "b" String somewhere in the grid, then return true, otherwise, return false.
public int shoot(int row, int col) - If row and col specify a square which is out of bounds, the method should return -1. If the square at the specified row and column contains "-" (i.e. is blank), the square should be changed to "m" to signify a miss, and the method should return 0. If the square contains "b" (i.e. a battleship which hasn't been hit yet) this square should be changed to an "x" to signify a hit, and the method should return 1. If the square contains either "x" or "m" the method should return 2 (these are squares which have already been “shot”).
public boolean gameOver() - Returns true only when the board contains no un-sunken ships. Otherwise, return false.
Along with your Board class there is another class in a separate file named Battleship. This class is a simple implementation of a game which allows you to add ships to a single Battleship board using user input and shoot at these ships. A sample run of the main method from this class is shown below.